+++ firebase-dart-admin-auth-sdk/example/firebase-dart-admin-auth-sdk-flutter-web-app/lib/screens/set_language_code_screen/set_language_code_screen.dart
@@ -5,7 +5,9 @@ import 'package:firebase/utils/extensions.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
+/// A screen for setting the language code.
 class SetLanguageCodeScreen extends StatefulWidget {
+  /// Constructs the [SetLanguageCodeScreen] widget.
   const SetLanguageCodeScreen({super.key});
 
   @override