+++ firebase-dart-admin-auth-sdk/example/firebase-dart-admin-auth-sdk-flutter-web-app/lib/shared/sign_in_methods_bottom_sheet.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: use_build_context_synchronouslypublic_member_api_docs uri_does_not_exist undefined_class undefined_function undefined_identifier undefined_method undefined_getter creation_with_non_type extends_non_class super_formal_parameter_without_associated_named undefined_super_member override_on_non_overriding_member non_type_as_type_argument non_constant_list_element unchecked_use_of_nullable_value
+// ignore_for_file: use_build_context_synchronously, public_member_api_docs, uri_does_not_exist, undefined_class, undefined_function, undefined_identifier, undefined_method, undefined_getter, creation_with_non_type, extends_non_class, super_formal_parameter_without_associated_named, undefined_super_member, override_on_non_overriding_member, non_type_as_type_argument, non_constant_list_element, unchecked_use_of_nullable_value
 
 import 'package:firebase/screens/sign_in_with_credential/sign_in_with_credential.dart';
 import 'package:firebase/screens/sign_in_with_email_and_password_screen/sign_in_with_email_and_password_screen.dart';
@@ -12,6 +12,7 @@ import 'package:flutter/material.dart';
 import 'package:firebase/screens/apple_sign_in_screen/apple_sign_in_screen.dart';
 import 'package:firebase/screens/gcp_sign_in_screen/gcp_sign_in_screen.dart';
 
+/// Shows the bottom sheet containing sign-in methods.
 void showSignMethodsBottomSheet(BuildContext context) {
   showModalBottomSheet(
     context: context,
@@ -20,7 +21,9 @@ void showSignMethodsBottomSheet(BuildContext context) {
   );
 }
 
+/// A bottom sheet that displays available sign-in methods.
 class SignInMethodsBottomSheet extends StatelessWidget {
+  /// Creates a [SignInMethodsBottomSheet].
   const SignInMethodsBottomSheet({super.key});
 
   @override